Correção de erros e mudança da lógica de retorno dos jogos

+++ src/controllers/DefaultController.ts
@@ -4,8 +4,10 @@ type Object = {
 
 export default class DefaultController {
   protected async destroyObjectArray(objects: Object[]) {
-    for (const object of objects) {
-      object.delete();
+    if (objects.length > 0) {
+      for (const object of objects) {
+        object.delete();
+      }
     }
   }
 }
